+++ frontend/src/services/sellerProduct.ts
@@ -1,0 +1,26 @@
+import { axiosDel, axiosGet, axiosPost, axiosPut } from "./api";
+
+// 내 상품 조회
+export const getProductMine = (args: { page?: number; size?: number }) => {
+  const { page = 0, size = 10 } = args;
+  return axiosGet("/seller/products/mine", { page, size });
+};
+
+// 상품 등록
+export const createProduct = (data: any) =>
+  axiosPost("/seller/products", data, {
+    headers: {
+      "Content-Type": "multipart/form-data",
+    },
+  });
+
+// 상품 수정
+export const updateProduct = (formData: any, productId: any) =>
+  axiosPut(`/seller/products/${productId}`, formData, {
+    headers: {
+      "Content-Type": "multipart/form-data",
+    },
+  });
+
+// 상품 삭제
+export const deleteProduct = (id: string) => axiosDel(`/seller/products/${id}`);
